Revert "Release FE (2023-12-13) (#1630)"
This reverts commit 4199e6a450ab4182657fe351dbbb23083f3f4e74.

+++ agents/agents/guard/calls.go
@@ -47,7 +47,7 @@ func (g Guard) verifyState(ctx context.Context, state types.State, stateIndex in
 		submitFunc = func(transactor *bind.TransactOpts) (tx *ethTypes.Transaction, err error) {
 			tx, err = g.domains[state.Origin()].LightInbox().VerifyStateWithAttestation(
 				transactor,
-				uint8(stateIndex),
+				int64(stateIndex),
 				data.SnapshotPayload(),
 				data.AttestationPayload(),
 				data.AttestationSignature(),
@@ -58,7 +58,7 @@ func (g Guard) verifyState(ctx context.Context, state types.State, stateIndex in
 		submitFunc = func(transactor *bind.TransactOpts) (tx *ethTypes.Transaction, err error) {
 			tx, err = g.domains[state.Origin()].LightInbox().VerifyStateWithSnapshot(
 				transactor,
-				uint8(stateIndex),
+				int64(stateIndex),
 				data.SnapshotPayload(),
 				data.SnapshotSignature(),
 			)
@@ -85,9 +85,9 @@ func (g Guard) verifyState(ctx context.Context, state types.State, stateIndex in
 
 type stateReportContract interface {
 	// SubmitStateReportWithSnapshot reports to the inbox that a state within a snapshot is invalid.
-	SubmitStateReportWithSnapshot(transactor *bind.TransactOpts, stateIndex uint8, signature signer.Signature, snapPayload []byte, snapSignature []byte) (tx *ethTypes.Transaction, err error)
+	SubmitStateReportWithSnapshot(transactor *bind.TransactOpts, stateIndex int64, signature signer.Signature, snapPayload []byte, snapSignature []byte) (tx *ethTypes.Transaction, err error)
 	// SubmitStateReportWithAttestation submits a state report corresponding to an attesation for an invalid state.
-	SubmitStateReportWithAttestation(transactor *bind.TransactOpts, stateIndex uint8, signature signer.Signature, snapPayload, attPayload, attSignature []byte) (tx *ethTypes.Transaction, err error)
+	SubmitStateReportWithAttestation(transactor *bind.TransactOpts, stateIndex int64, signature signer.Signature, snapPayload, attPayload, attSignature []byte) (tx *ethTypes.Transaction, err error)
 }
 
 // submitStateReport submits a state report to the given chain, provided a snapshot or attestation.
@@ -108,7 +108,7 @@ func (g Guard) submitStateReport(ctx context.Context, chainID uint32, state type
 		submitFunc = func(transactor *bind.TransactOpts) (tx *ethTypes.Transaction, err error) {
 			tx, err = contract.SubmitStateReportWithAttestation(
 				transactor,
-				uint8(stateIndex),
+				int64(stateIndex),
 				srSignature,
 				data.SnapshotPayload(),
 				data.AttestationPayload(),
@@ -120,7 +120,7 @@ func (g Guard) submitStateReport(ctx context.Context, chainID uint32, state type
 		submitFunc = func(transactor *bind.TransactOpts) (tx *ethTypes.Transaction, err error) {
 			tx, err = contract.SubmitStateReportWithSnapshot(
 				transactor,
-				uint8(stateIndex),
+				int64(stateIndex),
 				srSignature,
 				data.SnapshotPayload(),
 				data.SnapshotSignature(),
